+++ flutter_badge_manager_android/example/test/startup_states_test.dart
@@ -2,8 +2,9 @@ import 'package:flutter/services.dart';
 import 'package:flutter_badge_manager_example/main.dart' as example_app;
 import 'package:flutter_test/flutter_test.dart';
 
+import 'badge_api_mock.dart';
+
 void main() {
-  const badgeChannel = MethodChannel('github.com/ziqq/flutter_badge_manager');
   const permissionChannel = MethodChannel(
     'flutter.baseflow.com/permissions/methods',
   );
@@ -15,6 +16,7 @@ void main() {
     TestWidgetsFlutterBinding.ensureInitialized();
     badgeCalls.clear();
     permissionCalls.clear();
+    setUpBadgeApiMock(badgeCalls: badgeCalls);
 
     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
         .setMockMethodCallHandler(permissionChannel, (call) async {
@@ -28,27 +30,12 @@ void main() {
               return null;
           }
         });
-
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
-        .setMockMethodCallHandler(badgeChannel, (call) async {
-          badgeCalls.add(call);
-          switch (call.method) {
-            case 'isSupported':
-              return true;
-            case 'update':
-            case 'remove':
-              return null;
-            default:
-              throw PlatformException(code: 'unimplemented');
-          }
-        });
   });
 
   tearDown(() {
     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
         .setMockMethodCallHandler(permissionChannel, null);
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
-        .setMockMethodCallHandler(badgeChannel, null);
+    tearDownBadgeApiMock();
   });
 
   group('App startup -', () {